news_summary_app/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
schedule==1.2.0
python-dotenv==1.0.0
openai==1.3.3
email-validator==2.1.0
newspaper3k==0.2.8
lxml==4.9.3
nltk==3.8.1
python-dateutil==2.8.2